// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  /**
   * Supabase Storage URL
   */
  url?: string | null;
  /**
   * Path in Supabase Storage bucket
   */
  supabasePath?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  title: string;
  slug: string;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: number;
  title: string;
  slug: string;
  price: number;
  category: number | Category;
  images?: (number | Media)[] | null;
  shortDescription?: string | null;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  sku?: string | null;
  inventory?: number | null;
  status?: ('draft' | 'published') | null;
  updatedAt: string;
  createdAt: string;
}